refactor(lib, hicn-light, vpp, hiperf): HICN-723  - move infra data structure into the shared lib  - new packet cache using double hashing and lookup on prefix suffix  - testing updates  - authenticated requests using interest manifests

+++ hicn-plugin/src/interest_hitpit.h
@@ -46,7 +46,7 @@ typedef struct
 {
   u32 next_index;
   u32 sw_if_index;
-  u8 pkt_type;
+  u32 pkt_type;
 } hicn_interest_hitpit_trace_t;
 
 typedef enum
